refactor bluezutils to tools
I think this is a better name for the file.
Also changed code to call this file consistently in all files
