+++ examples/eddystone-url-beacon.py
@@ -4,12 +4,12 @@ import os
 sys.path.insert(0,
                 os.path.split(os.path.dirname(os.path.realpath(__file__)))[0])
 from bluezero import peripheral
-import bluezero.bluezutils as bluezutils
+from bluezero import tools
 
 # Bluetooth
 # Add beacon information
 ukBaz_beacon = peripheral.Service('FEAA', True, type='broadcast')
-service_data = bluezutils.url_to_advert(
+service_data = tools.url_to_advert(
     'https://github.com/ukBaz',
     frame_type=0x10,
     tx_power=0xFF)
